Fixed controller inheritance, reorganized get token code under yelp_helper, added sample code for iterating through yelp response data
