+++ app/controllers/home_controller.rb
@@ -1,4 +1,4 @@
-class HomeController < ActionController::Base
+class HomeController < ApplicationController
 
   def index
   end
