+++ app/controllers/route_list_controller.rb
@@ -1,7 +1,7 @@
 require_relative '../../lib/metro_rail'
 require_relative '../../lib/database_mgr'
 
-class RouteListController < ActionController::Base
+class RouteListController < ApplicationController
   @routelist = {}
 
   def show
